Add starting points on every specified steps
Now, `add_all_loci` gets a parameter `step` which specifies which loci
should be added. With step=1, all loci in a node will be added to
list of starting points. With any positive value of step=s every 's'
loci would be added to starting points list (each two loci in a node has
a distance of 's').

This value can be specified by command-line argument `-e` or
`--start-every`. The default value is 1 means adding all loci in a node.