Add `logger` header file
The `logger.h` is a wrapper for logging library (EasyLogging++). The
required macros were moved from `base.h` to this header file. This make
the library easily configurable through macro definitions and these
macros are bonded to the library header file. Thus, there is no need to
include `base.h` whenever logging library header file is included. It is
safer and less error-prone.