Add Haplotyper graph iterator
This graph iterator walk through the graph from a given start node to
the end (node that does not have any forward edge) while keeping track
of branch child nodes so that by running multiple times it can generate
unique simulated haplotype. The uniqueness is not guaranteed since the
traversal algorithm is greedy.