Change some graph iterator configurations
The `TSet` and `TContainer` types of BackTracker and Haplotyper,
respectively, converted from a basic type (Value) to the container
(vector) of the same type, since the TSet and TContainer are container
types. This change is required whenever it is needed to get the value
type of the set using `TSet::value_type`.