Use SeqAn's Seed class for seed hits
instead of custom structure. The 'Horizontal' string is used for
reference sequence set or graph, and the 'Vertical' string is used for
reads. The begin position of reference is sequence number or node id in
sequence and graph respectively, and its end position indicates the hit
position in the sequence or offset in the node specified by begin
position. For reads, the begin position is the read number in the string
set and the end position is the location in the read that the seed hits.