New Haplotyper traversal with better uniqueness
In order to generate a new unique path, new approach chooses the next
node (by calling `operator++` method) by looking at the visiting path
of length s (`setback` length) so far and check if it occured in the
already selected paths or not. In other word, assume that we are
generating path i, and the j number of nodes are already selected for
i-th path. For choosing the j+1 node, it iterates over all possible
(visited or unvisited) adjacent nodes -- let say a -- and forms a
setback path containing the nodes < Pi(j-s+2), ..., Pi(j), a >; where
P(k) is the k-th selected node of path i. If the setback path is not
covered by any previous paths, a is selected; otherwise it continues to
search for proper adjacent node. If all setback paths are already
covered, least covered node is selected. If all nodes are covered
equally, it uses a random adjacent.