Add another constructor for Traverser classes
At the time of instantiation of a Traverser, the starting point may not
be determined. So, another constructor is added to initialize the class
without specifying the starting locus. Since the model is that using
one Traverser for multiple loci in the graph, it would be more
convenient if it is instantiate without giving starting locus and then
set it by calling `set_start_locus()` method.