Add Path template class
This template class represents a path in the graph. All other scattered
helper function were encapsulated in this class. This class has two
specialization: normal path and compact path. The latter drops the
ordered array of node IDs, and can be used as a data structure for
efficient node ID presence queries.