Add output option to write seed hits
The output file contains a list of fixed-length integers written
subsequently. The length of each integer in bytes are determined by the
`sizeof(std::size_t)`. So, it's platform-dependent format. The final
alignment file would be in GAM file which is a stream of protobuf
message of Alignment which is cross-platform.