Add get position functions for Records iterators
In order to get the position of a seed in a Records of strings (set of
strings), some interface functions are added. They get ID of a string
in the Records object where seed lies as well as the offset of the seed
in that string. The `get_position` interface function returns the pair
of < ID, offset >.