Add direction to the PathSet
The interface function to map the position in the PathSet to node ID
and offset in the graph is also added. It is a wrapper function to call
the same interface function on the Path object in the set. For Reversed
PathSet the real position in the forward sequence will be passed to the
Path object.