Migrate to XG graph instead of VG
The VarGraph class had aggregated a VG protobuf message. The VG data
structure is not an efficient one to query in terms of memory model.
So, it is decided to be replaced by XG graph by using it as a base
class. By the way, the program gets the both format as input graphs and
converts the VG format to XG internally.