Fix the issue of SDSL temporary files location
The default value for `cache_config` structure in SDSL library is
current working directory meaning that the temporary files are created
in the current working directory where the program is executed. Now,
it is set to the value of TMPDIR environment variable if set; otherwise
the current working directory is used as before.