Fix the issue in finding seeds on the paths
SeqAn `Dna5` and `Dna5Q` alphabet values are 0 to 4 for A, C, G, T, and
N, respectively. But SDSL alphabets are handled with conversion table
`char2comp`. Thus, SDSL works with ASCII codes of the bases and SeqAn
uses integer values. The access operator (`operator*`) of the iterator
of the `seqan::Segment` class returns
`seqan::SimpleType< unsigned char, Dna5Q_ >`. In SDSL `backward_search`
function this value will be casted to `t_csa::char_type` (aka
`unsigned char` or `uint8_t`) yielding integer value of the alphabet;
where the SDSL expects the ASCII value. However, casting to `char` type
will yield the correct ASCII value of the alphabet. So, here the
`backward_search` function in the Finder class does exactly the same
algorithm that the SDSL `backward_search` does while it cast the
iterator value to `char` instead of `unsigned char`. That fixes the
issue.