Fix zero-filled bit vector resize issue
The function `sdsl::bit_vector::set_int()` get a range [idx, idx+len-1]
and set all bits in that range to integer representation of the given
parameter. However, `len` cannot be larger than 64 bit and it was used
wrongly.

The `resize_zf` method is added, and it gets a bit vector and size and
will resize the bit vector to new size and set the newly allocated bits
to zero using `memset` function if new size is larger than old size.