Fix the issue when TMPDIR is not set
Before, the `get_tmpdir` would return the value of environment variable
`TMPDIR`. If it was unset, it would return an empty string. In order to
avoid confusion, it is now renamed to `get_tmpdir_env` and the name is
used for another helper function (replacing the previous `get_tmpdir`)
that returns `GREM_DEFAULT_TMPDIR` which is set to '/tmp' by default if
`TMPDIR` is not set.