Remove nodes set attribute from Paths
The `std::set` data structure requires 30 Bytes for each 64-bit integer
entry in the set. So, for 200 million node IDs (each is a 64-bit
integer), apart from 1.6 GB for the data itself, it takes 12 GB extra
memory for auxilary data structure for constant-time data access. Thus,
for 16 paths it requires 224 GB and for 32 paths, 450 GB of memory which
we cannot afford for path index.

The nodes sets provided constant-time existance check for a node ID. By
removing this data structure from Path classes, querying a node ID to
check whether it is covered by a path is no longer constant-time. It is
linear to the number of nodes in the path. Consequently, choosing
"contains" strategy has no meaning now, because node ID availability
check for non-Micro paths cannot be done by using "Unordered" strategy.
For this reason, in addition to nodes set removal, the corresponding
interface functions `contains` and `covered_by` were adobted to this
change.

By removing the nodes set, the `sort`, `compress`, and the algorithm
used for `covered_by` is no longer meaningful or possible; since they
were using the nodes set for finding the node IDs with min/max values
(because the set was ordered) except for the `compress`. In addition,
`sort` does not improve the running time considerably. And `compress`
was used for decreasing number of paths in the paths vector. Now, with
more efficient way for node availability query, it is no longer needed
too.

This commit also fixes minor issues:

- Add `graph_type` typedef for `Path` template class.
- Make interface functions for Micro paths `inline`.
- Fix `contains` interface function friendship issue for Micro path.