Add PathSet template class
It represents a set of paths with efficient way to query node ID
presence in the paths set. It replaces the vector of paths in
`PathIndex` class.

In addition, the `PathIndex` is also polished to be consistent with the
rest of the code.

New `covered_by` interface function is added by using the path finding
provided by `PathSet`.