Fix zero-fill resize issue in sequence classes
The `resize_zf` would corrupts the memory allocated by
`sdsl::bit_vector` because it would replace a few bits beyond the
vector when the size of the bit vector was not byte-aligned. That would
raise SIGABRT when `sdsl::bit_vector` tries to deallocate the memory.

Now, the `resize_zf` is completely removed and the strategy has
changed: a bit vector with new size will be created. The content of the
old one is copied to the new one. Then they will be swapped. In order
to make the copy faster, the `bv_icopy` is added. It uses the `set_int`
and `get_int` member functions of `bit_vector` class (a.k.a
`int_vector`) to do the copy operation word-wise (64-bit).