Add non-const access to paths in PathSet
The `covered_by` interface function and other ones with node ID query
requires to create index if it has not already created. So, the PathSet
passed to the function could not be `const`. This only can possible if
PathSet could provide the non-const access function (`operator[]`, and
`begin` and `end` member functions).

Also, the PathSet passed to `add_all_loci` should also be `non-const`.