Fix moving index and its text fibre issue
The index has a pointer to the text fibre which it might not own it. It
is valid until the text object lives. When a class aggregates an index
(particularly, FMIndex) and its text fibre when an object of that class
is moved to another one. The text fibre of new object's index is no
longer valid since it points to the text fibre copy of the moved object
not the new one.

This commit is a work-around for this issue. The true solution would be
wrapping the text fibre in a Holder class which is responsible for the
memory management of the text fibre not the class aggregating the index
and its text. This Holder class can be developed from scratch or use
smart pointers in C++ Standard Library.