Fix 'init' rule issue
The master Makefile had an issue with Bear tools because of git
submodule update command while it is not needed at the first place. In
fact, the Makefile assumes that the distributed package contains all
necessary files, particularly fully populated `ext/` directory. If this
project is going to be build via git repository, the submodules should
be updated before attempting to build by user (this is already mentioned
in the README file).