Add Haplotype path
This path stores the nodes IDs as a bitvector of the size of number of
nodes in the graph. It may not be efficient representation of any path
in the graph, but it can be efficient for haplotype paths whose number
of nodes are O(n); where n is the number of nodes in the graph. Since
the bitvector just indicates the presence of the node IDs in the path,
this type of path is only applicable if the graph is DAG and the node
IDs are a topological sort of the graph. Since the path existence query
is efficient, the Haplotype path would be an efficient data structure
for implementing the Haplotyper iterator.