Add ggsim
The `ggsim` is a tool for simulating reads or haplotypes from a
variation graph. There is a key difference between `ggsim` and `vg sim`
in reads simulation which is the fact that `ggsim` generates the reads
from a number of haplotypes which are specified by ploidy of the genome
rather than simulating them from all paths of the underlying genome graph.