Add a tool for path index inspection
The `pathindex_stats` gets the path index prefix and the variation graph
and other required parameters. Then, it generates a vg Alignment file
containing all haplotypes in the path index along with all starting loci
which are marked as `S` base substitution (if `-L` is not set). It is
really useful for visualize the path index using SequenceTubeMap.