Fix Global Haplotyper iterator
This commit improves Haplotype selection algorithm. Instead of fixing
setback length, the length will be determined by the out-degree of
nodes as the Haplotype paths grows. The next node to append to the
Haplotype path is the node that its corresponding tailing sub-path does
not occur in the previously selected haplotyeps. The length of the
tailing sub-path is maximum number of nodes of the end of current
Haplotype path whose multiplication of their out-degrees are less than
the number of previously selected haplotypes.