Merge NamedStringSet to Records
Records class represents a set of sequence records including their name
and quality scores. Instead of vector of KSeq records, it decomposes
the set into separate vectors of names, comments, sequences, and quals.
This is useful when it is required to create an index over sequences
needed to be stored in one place all together.

But Records' fields (name and sequence) are aggregated into class
`NamedStringSet` which should be included in Records of Owner StringSet
class. Since it is only used as a base class for Records, it was dropped
out in this commit.