Fix reads chunks loading
Reading all reads from the file and loading it to the memory is not a
good idea, since they occupies substantial amount of memory. On the
other hand, profiling shows that using Dependent StringSet dramatically
reduces the performance of accessing each sequence. Besides, since it
used reads chunk data structure (Dependent StringSet) after being moved
to the Mapper object which may cause undefined behaviour. So, there are
lots of issues in this regard which this commit fixes it.

The SeqAn's sequence IO library is no longer used for parsing input
FASTQ file. It is migrated to "kseq++" library.