Fix YaPair class
The `i1`, and `i2` member variable of an instance of YaPair was
immutable before this change. Once, it is constructed the values cannot
be updated. That was a poor design.

This commit change those member variable to a reference to the
corresponding member variable of `std::pair` making the class working
with both `std`-style and `seqan`-style interfaces. The above-mentioned
issue is also resolved.

The corresponding test cases are added to the test scenarios of sequence
header file.