Better version and last update in the help message
In order to indicate which binary compiled from which version of the
code, it is necessary to include these information in the help message
of the program.

The output of `git describe` is used as more accurate version rather
than version in `release.h` header file. The former specifies if a build
is from dirty (untagged) version of the code.

The last update date in the help message was the the date of compilation
which does not mean anything. Now, it is updated to be the date of the
HEAD commit on which the program is built.