Check context size outside of the Mapper
Making the decision of the default context size in case it is not
specified or set to zero is a decision need to be done in
application-level rather than in Mapper module.

Apart from that, the program will check the context to be larger than
specified seed length making the program more robust in terms of input
parameter settings.