Hotfix: full path index loading issue
The full path index cannot be loaded when the specified context is
non-zero. This is because a non-zero context value is verified against
the stored value in the path index which is always zero. So, the `load`
method returns false and the path index is re-constructed which is not
desired.

In addition, this fact that seed length is smaller than context should
be verified after loading or constructing path index.