Improve `TimePeriod` class
This commit defines `duration`, `rep`, and `str` member function for `TimePeriod` class
to compute different measured time representations. Before that, it should be passed to
a static function of corresponding timer class which makes usage complicated.

Another change is using direct reference to stats member of the finder instead of
calling `get_instance_const_ptr`. The latter returns a constant pointer to the stats of
the tracking finder not the stats of the working one.