Distinguish between size type and ordinal type
The ordinal type can be much smaller than the size type (the type that can represent
the number of non-zero values). Distinct types for ordinal and size makes the data
structure more dynamic in terms of memory requirements. For example, the matrix
dimension for whole human genome can be fitted in 32-bit integer, while the number of
non-zero values can be represented by not smaller than 64-bit integer. In this case, we
can have an entries array of 32-bit integers instead of an array of 64-bit.