Report the end of reverse hits
PSI only sees graphs as forward. So, seed hits are reported by their start of the
occurrence in both the graph and the read. This commit enables caller to report the end
of hit occurrences for reads with odd IDs (considered to be reverse complement) in full
sensitive mode or when `reversed` parameter is `true` when finding MEMs.

However, for `GraphAligner`, it does not matter if the start of a hit is reported or the
end of it as long as it is consistent with each other. Reverse hits should be reported
by its position on the reversed graph (either end or start) and its position on the
forward sequence of the read (either end or start consistent with the position in the
graph).

This commit is not needed anymore, but it has been kept in a side branch.