More control on random number engine for tests
First, a separated thread local random number engine is used for test
cases. The seed will be set via `--rng-seed` argument by a custom event
listener which will be called once when testing is started. When a test
fails, in order to reproduce the test data, the value of seed can be
reported by `INFO( rnd::iseed )`.