Change vg::Position to psi::Position
This invalidates starting loci files in PSI index as the serialisation
format has changed.

Since there is not any specialisation of template function `serialize`
for `RandomAccessProxyContainer`, more explicit specialisation for
containers has been used by passing the start and end iterators to
indicate that the input argument `sloci` is a container.

[squash] de/serialise starts