Sign conversion in Compact path assignments
The `value_type` of containers in all path specialisations, but Compact,
are `id_type`. The node ids container in Compact path is `enc_vector`
which assumes the input container's `value_type` is unsigned. While the
node ID type, `id_type`, is signed integer. Assigning a signed vector
to `enc_vector` raised many sign-compare warnings.

In this commit, a new `assign` function overload is defined which
converts each node id to the counterpart unsigned type if the
`value_type` of the input container is signed.