Decouple `adjacency_matrix` from PairG
Instead of relying on types defined on `pairg` namespace, the function
get matrix type itself as a template parameter and deduce other types
from it. The order of template parameters has also changed and the
output type should be provided explicitly as a template parameter. The
graph type is deduced from the graph argument.