Refactor HBitVector
Fixes several bugs and improve the implementation, including:

- Fixed SIGSEGV error which was caused by negative relative ids. Adding
the size of the vector guarantees that the result of the modulo is
always positive.

- It also fixes a bug in setting bits when both ends of a range occur on
a single bitset. Apparently, shifting more than the width of the bitset
has an undefined behaviour according to C++11 (from Section 5.8 p1 in
the C++11 draft).

- Fix data race

- Fix relative index calculation: If the size of the bitvector is not a
multiple of bitset width, the last bitset in L2 will be shared between
low-L2 and high-L2 which make it difficult to convert set bits back to
intervals. This commit fixes this issue by considering the "aligned"
size instead of the actual size which is the smallest multiple of bitset
width that can accommodate the vector. So, internally the bitvector
assumes that all bits in allocated aligned memory is addressable,
although the caller may never access them.

- This commit also polishes the code, defines const methods, provides
better calculation of L1/L2 sizes and number of bitsets. It also removes
the `m_size` as it is unused. In this commit, static methods are not
called by `this` pointer anymore.